116 Beaconside
"That Miss Flora Millar, the lady who had caused the disturbance, has actually been arrested. It
appears that she was formerly a danseuse at the Allegro, and that she has known the bridegroom for
some years. There are no further particulars, and the whole case is in your hands now--so far as it
has been set forth in the public press."

"And an exceedingly interesting case it appears to be. I would not have missed it for worlds. But
there is a ring at the bell, Watson, and as the clock makes it a few minutes after four, I have no
doubt that this will prove to be our noble client. Do not dream of going, Watson, for I very much
prefer having a witness, if only as a check to my own memory."

"Lord Robert St. Simon," announced our page-boy, throwing open the door. A gentleman entered, with a
pleasant, cultured face, high-nosed and pale, with something perhaps of petulance about the mouth,
and with the steady, well-opened eye of a man whose pleasant lot it had ever been to command and to
be obeyed. His manner was brisk, and yet his general appearance gave an undue impression of age, for
he had a slight forward stoop and a little bend of the knees as he walked. His hair, too, as he
swept off his very curly-brimmed hat, was grizzled round the edges and thin upon the top. As to his
dress, it was careful to the verge of foppishness, with high collar, black frock-coat, white
waistcoat, yellow gloves, patent-leather shoes, and light-coloured gaiters. He advanced slowly into
the room, turning his head from left to right, and swinging in his right hand the cord which held
his golden eyeglasses.

"Good-day, Lord St. Simon," said Holmes, rising and bowing. "Pray take the basket-chair. This is my
friend and colleague, Dr. Watson. Draw up a little to the fire, and we will talk this matter over."