108 Harvey Street
"Nothing?"

"To tell the truth"--he sank his face into his thin, white hands--"I have felt helpless. I have felt
like one of those poor rabbits when the snake is writhing towards it. I seem to be in the grasp of
some resistless, inexorable evil, which no foresight and no precautions can guard against."

"Tut! tut!" cried Sherlock Holmes. "You must act, man, or you are lost. Nothing but energy can save
you. This is no time for despair."

"I have seen the police."

"Ah!"

"But they listened to my story with a smile. I am convinced that the inspector has formed the
opinion that the letters are all practical jokes, and that the deaths of my relations were really
accidents, as the jury stated, and were not to be connected with the warnings."