29 Valentia Road
How shall I ever forget that dreadful vigil? I could not hear a sound, not even the drawing of a
breath, and yet I knew that my companion sat open-eyed, within a few feet of me, in the same state
of nervous tension in which I was myself. The shutters cut off the least ray of light, and we waited
in absolute darkness.

From outside came the occasional cry of a night-bird, and once at our very window a long drawn
catlike whine, which told us that the cheetah was indeed at liberty. Far away we could hear the deep
tones of the parish clock, which boomed out every quarter of an hour. How long they seemed, those
quarters! Twelve struck, and one and two and three, and still we sat waiting silently for whatever
might befall.