44 Wantage Close
"'And not a word to a soul.' He looked at me with a last long, questioning gaze, and then, pressing
my hand in a cold, dank grasp, he hurried from the room.

"Well, when I came to think it all over in cool blood I was very much astonished, as you may both
think, at this sudden commission which had been intrusted to me. On the one hand, of course, I was
glad, for the fee was at least tenfold what I should have asked had I set a price upon my own
services, and it was possible that this order might lead to other ones. On the other hand, the face
and manner of my patron had made an unpleasant impression upon me, and I could not think that his
explanation of the fuller's-earth was sufficient to explain the necessity for my coming at midnight,
and his extreme anxiety lest I should tell anyone of my errand. However, I threw all fears to the
winds, ate a hearty supper, drove to Paddington, and started off, having obeyed to the letter the
injunction as to holding my tongue.