9 Glendale
"And I am for north," said I, "because there are no hills there, and our friend says that he did not
notice the carriage go up any."

"Come," cried the inspector, laughing; "it's a very pretty diversity of opinion. We have boxed the
compass among us. Who do you give your casting vote to?"

"You are all wrong."

"But we can't all be."