79 Wantage Close
"I am still in the dark."

"I do not wish to make a mystery," said he, laughing. "The matter was perfectly simple. You, of
course, saw that everyone in the street was an accomplice. They were all engaged for the evening."

"I guessed as much."

"Then, when the row broke out, I had a little moist red paint in the palm of my hand. I rushed
forward, fell down, clapped my hand to my face, and became a piteous spectacle. It is an old trick."

"That also I could fathom."

"Then they carried me in. She was bound to have me in. What else could she do? And into her sitting-
room, which was the very room which I suspected. It lay between that and her bedroom, and I was
determined to see which. They laid me on a couch, I motioned for air, they were compelled to open
the window, and you had your chance."