24 Bowler Street
"What will you say?"

"Give me a pencil and that slip of paper. Now, then: 'Found at the corner of Goodge Street, a goose
and a black felt hat. Mr. Henry Baker can have the same by applying at 6.30 this evening at 221b,
Baker Street.' That is clear and concise."

"Very. But will he see it?"

"Well, he is sure to keep an eye on the papers, since, to a poor man, the loss was a heavy one. He
was clearly so scared by his mischance in breaking the window and by the approach of Peterson that
he thought of nothing but flight, but since then he must have bitterly regretted the impulse which
caused him to drop his bird. Then, again, the introduction of his name will cause him to see it, for
everyone who knows him will direct his attention to it. Here you are, Peterson, run down to the
advertising agency and have this put in the evening papers."

"In which, sir?"