60 Moor End
"I never hope to see such a sight as that again, Mr. Holmes. From north, south, east, and west every
man who had a shade of red in his hair had tramped into the city to answer the advertisement. Fleet
Street was choked with red-headed folk, and Pope's Court looked like a coster's orange barrow. I
should not have thought there were so many in the whole country as were brought together by that
single advertisement. Every shade of colour they were--straw, lemon, orange, brick, Irish-setter,
liver, clay; but, as Spaulding said, there were not many who had the real vivid flame-coloured tint.
When I saw how many were waiting, I would have given it up in despair; but Spaulding would not hear
of it. How he did it I could not imagine, but he pushed and pulled and butted until he got me
through the crowd, and right up to the steps which led to the office. There was a double stream upon
the stair, some going up in hope, and some coming back dejected; but we wedged in as well as we
could and soon found ourselves in the office."

"Your experience has been a most entertaining one," remarked Holmes as his client paused and
refreshed his memory with a huge pinch of snuff. "Pray continue your very interesting statement."

"There was nothing in the office but a couple of wooden chairs and a deal table, behind which sat a
small man with a head that was even redder than mine. He said a few words to each candidate as he
came up, and then he always managed to find some fault in them which would disqualify them. Getting
a vacancy did not seem to be such a very easy matter, after all. However, when our turn came the
little man was much more favourable to me than to any of the others, and he closed the door as we
entered, so that he might have a private word with us.

"'This is Mr. Jabez Wilson,' said my assistant, 'and he is willing to fill a vacancy in the League.'

"'And he is admirably suited for it,' the other answered. 'He has every requirement. I cannot recall
when I have seen anything so fine.' He took a step backward, cocked his head on one side, and gazed
at my hair until I felt quite bashful. Then suddenly he plunged forward, wrung my hand, and
congratulated me warmly on my success.