6 Sunningdale Drive
The Count shrugged his broad shoulders. "Then I must begin," said he, "by binding you both to
absolute secrecy for two years; at the end of that time the matter will be of no importance. At
present it is not too much to say that it is of such weight it may have an influence upon European
history."

"I promise," said Holmes.

"And I."