105 Sunningdale Drive
"Your cases have indeed been of the greatest interest to me," I observed.

"You will remember that I remarked the other day, just before we went into the very simple problem
presented by Miss Mary Sutherland, that for strange effects and extraordinary combinations we must
go to life itself, which is always far more daring than any effort of the imagination."

"A proposition which I took the liberty of doubting."

"You did, Doctor, but none the less you must come round to my view, for otherwise I shall keep on
piling fact upon fact on you until your reason breaks down under them and acknowledges me to be
right. Now, Mr. Jabez Wilson here has been good enough to call upon me this morning, and to begin a
narrative which promises to be one of the most singular which I have listened to for some time. You
have heard me remark that the strangest and most unique things are very often connected not with the
larger but with the smaller crimes, and occasionally, indeed, where there is room for doubt whether
any positive crime has been committed. As far as I have heard it is impossible for me to say whether
the present case is an instance of crime or not, but the course of events is certainly among the
most singular that I have ever listened to. Perhaps, Mr. Wilson, you would have the great kindness
to recommence your narrative. I ask you not merely because my friend Dr. Watson has not heard the
opening part but also because the peculiar nature of the story makes me anxious to have every
possible detail from your lips. As a rule, when I have heard some slight indication of the course of
events, I am able to guide myself by the thousands of other similar cases which occur to my memory.
In the present instance I am forced to admit that the facts are, to the best of my belief, unique."