37 Glan Road
"What can be the matter, then? Can you suggest no explanation?"

"I have devised seven separate explanations, each of which would cover the facts as far as we know
them. But which of these is correct can only be determined by the fresh information which we shall
no doubt find waiting for us. Well, there is the tower of the cathedral, and we shall soon learn all
that Miss Hunter has to tell."

The Black Swan is an inn of repute in the High Street, at no distance from the station, and there we
found the young lady waiting for us. She had engaged a sitting-room, and our lunch awaited us upon
the table.

"I am so delighted that you have come," she said earnestly. "It is so very kind of you both; but
indeed I do not know what I should do. Your advice will be altogether invaluable to me."

"Pray tell us what has happened to you."

"I will do so, and I must be quick, for I have promised Mr. Rucastle to be back before three. I got
his leave to come into town this morning, though he little knew for what purpose."