2 Badgers Dene
"Oh, yes, sir; I have only just left him."

"And has your business been attended to in your absence?"

"Nothing to complain of, sir. There's never very much to do of a morning."

"That will do, Mr. Wilson. I shall be happy to give you an opinion upon the subject in the course of
a day or two. To-day is Saturday, and I hope that by Monday we may come to a conclusion."

"Well, Watson," said Holmes when our visitor had left us, "what do you make of it all?"