64 Wilson Gardens
With a dazed face the banker made out the required check. Holmes walked over to his desk, took out a
little triangular piece of gold with three gems in it, and threw it down upon the table.

With a shriek of joy our client clutched it up.

"You have it!" he gasped. "I am saved! I am saved!"

The reaction of joy was as passionate as his grief had been, and he hugged his recovered gems to his
bosom.