30 Sunningdale Drive
"It is a murder, then?"

"Well, it is conjectured to be so. I shall take nothing for granted until I have the opportunity of
looking personally into it. I will explain the state of things to you, as far as I have been able to
understand it, in a very few words.