58 Glan Road
"Is Briony Lodge, Serpentine Avenue, St. John's Wood."

Holmes took a note of it. "One other question," said he. "Was the photograph a cabinet?"

"It was."

"Then, good-night, your Majesty, and I trust that we shall soon have some good news for you. And
good-night, Watson," he added, as the wheels of the royal brougham rolled down the street. "If you
will be good enough to call to-morrow afternoon at three o'clock I should like to chat this little
matter over with you."

At three o'clock precisely I was at Baker Street, but Holmes had not yet returned. The landlady
informed me that he had left the house shortly after eight o'clock in the morning. I sat down beside
the fire, however, with the intention of awaiting him, however long he might be. I was already
deeply interested in his inquiry, for, though it was surrounded by none of the grim and strange
features which were associated with the two crimes which I have already recorded, still, the nature
of the case and the exalted station of his client gave it a character of its own. Indeed, apart from
the nature of the investigation which my friend had on hand, there was something in his masterly
grasp of a situation, and his keen, incisive reasoning, which made it a pleasure to me to study his
system of work, and to follow the quick, subtle methods by which he disentangled the most
inextricable mysteries. So accustomed was I to his invariable success that the very possibility of
his failing had ceased to enter into my head.

It was close upon four before the door opened, and a drunken-looking groom, ill-kempt and side-
whiskered, with an inflamed face and disreputable clothes, walked into the room. Accustomed as I was
to my friend's amazing powers in the use of disguises, I had to look three times before I was
certain that it was indeed he. With a nod he vanished into the bedroom, whence he emerged in five
minutes tweed-suited and respectable, as of old. Putting his hands into his pockets, he stretched
out his legs in front of the fire and laughed heartily for some minutes.