2 Balcombe Close
"He's a brave fellow," said a woman. "They would have had the lady's purse and watch if it hadn't
been for him. They were a gang, and a rough one, too. Ah, he's breathing now."

"He can't lie in the street. May we bring him in, marm?"

"Surely. Bring him into the sitting-room. There is a comfortable sofa. This way, please!"

Slowly and solemnly he was borne into Briony Lodge and laid out in the principal room, while I still
observed the proceedings from my post by the window. The lamps had been lit, but the blinds had not
been drawn, so that I could see Holmes as he lay upon the couch. I do not know whether he was seized
with compunction at that moment for the part he was playing, but I know that I never felt more
heartily ashamed of myself in my life than when I saw the beautiful creature against whom I was
conspiring, or the grace and kindliness with which she waited upon the injured man. And yet it would
be the blackest treachery to Holmes to draw back now from the part which he had intrusted to me. I
hardened my heart, and took the smoke-rocket from under my ulster. After all, I thought, we are not
injuring her. We are but preventing her from injuring another.

Holmes had sat up upon the couch, and I saw him motion like a man who is in need of air. A maid
rushed across and threw open the window. At the same instant I saw him raise his hand and at the
signal I tossed my rocket into the room with a cry of "Fire!" The word was no sooner out of my mouth
than the whole crowd of spectators, well dressed and ill--gentlemen, ostlers, and servant-maids--
joined in a general shriek of "Fire!" Thick clouds of smoke curled through the room and out at the
open window. I caught a glimpse of rushing figures, and a moment later the voice of Holmes from
within assuring them that it was a false alarm. Slipping through the shouting crowd I made my way to
the corner of the street, and in ten minutes was rejoiced to find my friend's arm in mine, and to
get away from the scene of uproar. He walked swiftly and in silence for some few minutes until we
had turned down one of the quiet streets which lead towards the Edgeware Road.