150 Wantage Close
"'It's only four hours a day, and I should not think of leaving,' said I.

"'No excuse will avail,' said Mr. Duncan Ross; 'neither sickness nor business nor anything else.
There you must stay, or you lose your billet.'

"'And the work?'

"'Is to copy out the "Encyclopaedia Britannica." There is the first volume of it in that press. You
must find your own ink, pens, and blotting-paper, but we provide this table and chair. Will you be
ready to-morrow?'