115 Harvey Street
"I have seen those symptoms before," said Holmes, throwing his cigarette into the fire. "Oscillation
upon the pavement always means an affaire de coeur. She would like advice, but is not sure that the
matter is not too delicate for communication. And yet even here we may discriminate. When a woman
has been seriously wronged by a man she no longer oscillates, and the usual symptom is a broken bell
wire. Here we may take it that there is a love matter, but that the maiden is not so much angry as
perplexed, or grieved. But here she comes in person to resolve our doubts."

As he spoke there was a tap at the door, and the boy in buttons entered to announce Miss Mary
Sutherland, while the lady herself loomed behind his small black figure like a full-sailed merchant-
man behind a tiny pilot boat. Sherlock Holmes welcomed her with the easy courtesy for which he was
remarkable, and, having closed the door and bowed her into an armchair, he looked her over in the
minute and yet abstracted fashion which was peculiar to him.

"Do you not find," he said, "that with your short sight it is a little trying to do so much
typewriting?"

"I did at first," she answered, "but now I know where the letters are without looking." Then,
suddenly realising the full purport of his words, she gave a violent start and looked up, with fear
and astonishment upon her broad, good-humoured face. "You've heard about me, Mr. Holmes," she cried,
"else how could you know all that?"